refactor: App.tsx 코드 분리
