+++ src/navigators/TabNavigator.tsx
@@ -1,0 +1,48 @@
+import * as React from 'react';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { Ionicons } from '@expo/vector-icons';
+import HomeScreen from '../screens/home/HomeScreen';
+import ExploreScreen from '../screens/ExploreScreen';
+import ExampleScreen from '../screens/ExampleScreen';
+
+const Tab = createBottomTabNavigator();
+
+function TabNavigator() {
+  return (
+    <Tab.Navigator
+      screenOptions={({ route }) => ({
+        tabBarIcon: ({ focused, color, size }) => {
+          let iconName: keyof typeof Ionicons.glyphMap = 'home-outline';
+          if (route.name === 'Home') {
+            iconName = focused ? 'home' : 'home-outline';
+          } else if (route.name === 'Explore') {
+            iconName = focused ? 'search' : 'search-outline';
+          } else if (route.name === 'Example') {
+            iconName = focused ? 'star' : 'star-outline';
+          }
+          return <Ionicons name={iconName} size={size} color={color} />;
+        },
+        tabBarActiveTintColor: 'tomato',
+        tabBarInactiveTintColor: 'gray',
+      })}
+    >
+      <Tab.Screen
+        name="Home"
+        component={HomeScreen}
+        options={{ title: 'Do! Farming', headerShown: false }}
+      />
+      <Tab.Screen
+        name="Explore"
+        component={ExploreScreen}
+        options={{ title: 'Explore Screen' }}
+      />
+      <Tab.Screen
+        name="Example"
+        component={ExampleScreen}
+        options={{ title: 'Example Screen' }}
+      />
+    </Tab.Navigator>
+  );
+}
+
+export default TabNavigator;
